-i updated the discounted products UI ; -the review feature is 100% ready; - i converted the products reducer to redux slice; -i added mixed products section to home page;

+++ BackEnd/Solution/APIs_Layer/Controllers/ReviewController.cs
@@ -56,7 +56,7 @@ namespace APIs.Controllers
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        public async Task<IActionResult> GetReviewById(int id)
+        public async Task<ActionResult<ReadReviewDTO>> GetReviewById(int id)
         { 
             if (id <= 0) return BadRequest("invalid review-id");
             
@@ -91,12 +91,62 @@ namespace APIs.Controllers
             }
         }
 
-        // DELETE: api/review/{id}
+        [HttpPut("updateReview")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+
+        public async Task<ActionResult<bool>>UpdateReview(ReviewDTO review)
+        {
+            if (review == null ) return BadRequest("review is null");
+            if (review.Id <= 0 ) return BadRequest("invalid reviewId");
+            if (review.UserId <= 0 ) return BadRequest("invalid userId");
+            if (review.ProductId <= 0 ) return BadRequest("invalid productId");
+            if (string.IsNullOrEmpty(review.Comment) ) return BadRequest("comment is empty");
+            if (review.Rating <1 || review.Rating > 5) return BadRequest("rating range must be between 1 and 5 stars");
+
+            try
+            {
+                var UpdatedReview = new Review
+                {
+                    Id = review.Id,
+                    UserId = review.UserId,
+                    ProductId = review.ProductId,
+                    Comment = review.Comment,
+                    Rating = review.Rating
+
+                };
+               bool isUpdated =await _reviewService.UpdateReviewAsync(UpdatedReview);
+
+                return isUpdated? Ok(isUpdated) : NotFound("review not exists in the system");
+            }
+            catch(Exception ex)
+            {
+                return StatusCode(500, $"internal server error : [{ex.Message}]");
+            }
+
+        }
+
+
         [HttpDelete("{id}")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<IActionResult> DeleteReview(int id)
         {
-            var deleted = await _reviewService.DeleteReviewByIdAsync(id);
-            return deleted ? NoContent() : NotFound();
+            if (id <= 0) return BadRequest("invalid id");
+            try
+            {
+                var deleted = await _reviewService.DeleteReviewByIdAsync(id);
+                return deleted ? Ok("review has been deleted successfully") : NotFound("review not found");
+            }
+            catch(Exception ex)
+            {
+                return StatusCode(500, $"internal server error ({ex.Message})");
+            }
+           
         }
 
         [HttpGet("product/{productId}")]
